added transition animation when moving the cursor in the menu

+++ src/main/java/com/ccdr/labyrinth/menu/tree/MenuListElement.java
@@ -37,7 +37,7 @@ public class MenuListElement extends MenuElement{
         menuElement.setParent(this);
     }
 
-    public double getIndex() {
+    public int getIndex() {
         return this.index;
     }
